Renamed ReadTokenData folder.:
