+++ src/components/common/index.ts
@@ -2,9 +2,9 @@ export { Checkbox } from './Checkbox';
 export { CookiePopup } from './CookiePopup';
 export { CouldNotLoad } from './CouldNotLoad';
 export { Input } from './Input';
+export { LoadUserData } from './LoadUserData';
 export { LogoutPopup } from './LogoutPopup';
 export { Modal } from './Modal';
 export { PrivateRoute } from './PrivateRoute';
-export { LoadUserData } from './ReadTokenData';
 export * as Select from './Select';
 export { SEO } from './SEO';
